+++ kos/example/token_service/include/general.h
@@ -2,12 +2,13 @@
  * © 2024 AO Kaspersky Lab
  * Licensed under the MIT License
  */
+
 #ifndef TOKEN_SERVICE_INCLUDE_GENERAL_H
 #define TOKEN_SERVICE_INCLUDE_GENERAL_H
 #pragma once
 
 namespace app {
-constexpr auto AppTag = "TokenService";
+constexpr auto AppTag = "[TokenService] ";
 }  // namespace app
 
 #endif  // TOKEN_SERVICE_INCLUDE_GENERAL_H