+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Release for KasperskyOS Community Edition SDK 1.3 (2025-07)
+
+### Changes
+
+* Replaced deprecated word "entity" by "task class"
+* Updated STANDARD_GNU compiler flag to version 17
+* Added support for the Radxa ROCK 3A hardware platform
+* Added OpenSSL git submodule
+
 ## Initial adaptation for KasperskyOS - 2024-06-26
 
 ### Changes